refact app
  * rename Init() to InitPlayer()
  * move InitPlayer() to package internal/pkg/initPlayer
  * include package initPlayer via go.mod
