+++ go.mod
@@ -1,0 +1,7 @@
+module go-tic-tac-toe
+
+go 1.17
+
+replace go-tic-tac-toe/internal/pkg/initPlayer v0.0.0 => ./internal/pkg/initPlayer
+
+require go-tic-tac-toe/internal/pkg/initPlayer v0.0.0
