+++ internal/pkg/initPlayer/go.mod
@@ -1,0 +1,3 @@
+module go-tic-tac-toe/internal/pkg/initPlayer
+
+go 1.17
